Moved files to core folder
